pyproject.toml
[tool.poetry]
name = "journal"
version = "0.1.0"
description = ""
authors = ["Aratimius <aratimius@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg2-binary = "^2.9.10"
ipython = "^8.29.0"
python-dotenv = "^1.0.1"
black = "^24.10.0"
django-crispy-forms = "^2.3"
crispy-tailwind = "^1.0.3"
django = "4.2"
redis = "^5.2.0"
django-apscheduler = "^0.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
